Ok, forgot two files and worked out how to cast fields. For the moment just ints will do, tagging is next.

+++ part-3/3-2 RequestData in C++/main.cpp
@@ -32,7 +32,7 @@ struct AircraftInfo {
     std::string title;
     std::string tailNumber;
     std::string atcId;
-    double altitude;
+    int altitude;
     double latitude;
     double longitude;
 };
